moved adobe code into xds_adobe
